Refactor: Move function argument mapping code to a new module
Most of these are imported in more than one place (in mypy or
mypyc) so it makes sense to put them in a dedicated module.
